# docs/reference/openapi.yaml
openapi: 3.1.0
info:
  title: skelm gateway
  description: |
    Control plane and runs surface for the skelm gateway. The gateway is the
    trust boundary — every privileged action is enforced and audited here.
  version: 0.3.9
  license:
    name: MIT
servers:
  - url: http://127.0.0.1:14738
    description: Default local gateway
security:
  - bearerAuth: []
paths:
...
  /pipelines:
    get:
      summary: List registered pipelines
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/PipelineSummary' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /pipelines/{id}:
    get:
      summary: Describe one pipeline
      parameters:
        - $ref: '#/components/parameters/PipelineId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PipelineDescription' }
        '404': { $ref: '#/components/responses/NotFound' }
  /pipelines/{id}/run:
    post:
      summary: Run a pipeline synchronously
      parameters:
        - $ref: '#/components/parameters/PipelineId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                input: {}
      responses:
        '200':
          description: Run completed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Run' }
        '422':
          description: Schema validation failure
        '500':
          description: Run failed at runtime
  /pipelines/{id}/start:
    post:
      summary: Start an asynchronous run
      parameters:
        - $ref: '#/components/parameters/PipelineId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                input: {}
      responses:
        '202':
          description: Run accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  runId: { type: string }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  parameters:
    PipelineId:
      in: path
      name: id
      required: true
      schema: { type: string }
...
  responses:
    Unauthorized:
      description: Missing or invalid bearer token
    NotFound:
      description: Resource not found
  schemas:
    PipelineSummary:
      type: object
      properties:
        id: { type: string }
        version: { type: string }
    PipelineDescription:
      type: object
      properties:
        id: { type: string }
        steps:
          type: array
          items: { type: object }
    Run:
      type: object
      properties:
        runId: { type: string }
        pipelineId: { type: string }
        status:
          type: string
          enum: [queued, running, waiting, completed, failed, cancelled]
        output: {}
        error:
          type: object
          properties:
            message: { type: string }
            code: { type: string }
...